+++ src/app/components/embedded-content-file-browser.js
@@ -10,13 +10,14 @@
 import OneEmbeddedContainer from 'onezone-gui/components/one-embedded-container';
 import layout from 'onezone-gui/templates/components/one-embedded-container';
 import { inject as service } from '@ember/service';
-import { computed } from '@ember/object';
+import { computed, observer } from '@ember/object';
 import { reads } from '@ember/object/computed';
 
 export default OneEmbeddedContainer.extend({
   layout,
 
   globalNotify: service(),
+  router: service(),
 
   /**
    * @virtual
@@ -25,21 +26,23 @@ export default OneEmbeddedContainer.extend({
   oneprovider: undefined,
 
   /**
-   * TODO: to decide if using entity or record id
-   * ID of `file` record of directory to show in files browser.
-   * It can be changed from Onezone GUI to change dir displayed
-   * in Oneprovider's iframe.
-   * @type {string}
-   */
-  fileId: undefined,
-
-  /**
-   * TODO: to decide if using entity or record id
-   * ID of `space` record that is space of directory displayed in files
+   * Entity ID of `space` record that is space of directory displayed in files
    * browser.
    * @type {string}
    */
-  spaceId: undefined,
+  spaceEntityId: undefined,
+
+  /**
+   * Entity ID of `file` record that is the directory displayed in files
+   * browser.
+   * @type {string}
+   */
+  dirEntityId: computed('router', function dirEntityId() {
+    // FIXME: find a better way
+    const currentUrl = this.get('router').get('currentURL');
+    const m = currentUrl.match(/.*options=dir\.(.*)/);
+    return m && m[1];
+  }),
 
   /**
    * @override implements OneEmbeddedContainer
@@ -60,12 +63,12 @@ export default OneEmbeddedContainer.extend({
   /**
    * @override implements OneEmbeddedContainer
    */
-  iframeInjectedProperties: Object.freeze(['fileId', 'spaceId']),
+  iframeInjectedProperties: Object.freeze(['spaceEntityId', 'dirEntityId']),
 
   /**
    * @override implements OneEmbeddedContainer
    */
-  callParentActionNames: Object.freeze(['sayHello']),
+  callParentActionNames: Object.freeze(['updateDirEntityId']),
 
   /**
    * @override implements OneEmbeddedContainer
@@ -75,11 +78,22 @@ export default OneEmbeddedContainer.extend({
     return `iframe-oneprovider-${oneproviderId}`;
   }),
 
+  // TODO: if there will be more params, use some utils for changing options hash
+  dirChangedObserver: observer('dirEntityId', function dirChangedObserver() {
+    const {
+      router,
+      dirEntityId,
+    } = this.getProperties('router', 'dirEntityId');
+    router.transitionTo({
+      queryParams: {
+        options: `dir.${dirEntityId}`,
+      },
+    });
+  }),
+
   actions: {
-    // TODO: integration testing code, to remove when content-file-browser
-    // of Oneprovider GUI will be ready
-    sayHello() {
-      this.get('globalNotify').info('hello');
+    updateDirEntityId(dirEntityId) {
+      this.set('dirEntityId', dirEntityId);
     },
   },
 });