+++ src/app/locales/en/components/content-spaces-data.js
@@ -1,0 +1,4 @@
+export default {
+  allOneprovidersOffline: 'All supporting Oneproviders are offline',
+  oneproviderOffline: 'Selected Oneprovider is currently offline',
+};