+++ src/app/locales/en/components/cookies-consent.js
@@ -1,0 +1,4 @@
+export default {
+  header: 'This site uses cookies',
+  acceptButton: 'I understand',
+};