+++ src/app/locales/en/errors/backend-errors.js
@@ -1,0 +1,3 @@
+export default {
+  cannotConnectToElasticsearch: 'Cannot connect to Elasticsearch.',
+};