+++ src/app/models/provider.js
@@ -11,14 +11,12 @@ import Model from 'ember-data/model';
 import attr from 'ember-data/attr';
 import { computed } from '@ember/object';
 import { belongsTo } from 'onedata-gui-websocket-client/utils/relationships';
+import StaticGraphModelMixin from 'onedata-gui-websocket-client/mixins/models/static-graph-model';
 import GraphSingleModelMixin from 'onedata-gui-websocket-client/mixins/models/graph-single-model';
 import parseGri from 'onedata-gui-websocket-client/utils/parse-gri';
 
 export default Model.extend(GraphSingleModelMixin, {
   name: attr('string'),
-  // TODO: add array transform
-  // urls: attr('array'),
-  // clientName: attr('string'),
   longitude: attr('number', { defaultValue: 0 }),
   latitude: attr('number', { defaultValue: 0 }),
   online: attr('boolean'),
@@ -40,4 +38,4 @@ export default Model.extend(GraphSingleModelMixin, {
   }),
   //#endregion
 
-});
+}).reopenClass(StaticGraphModelMixin);