+++ src/app/utils/generate-development-model.js
@@ -9,10 +9,9 @@
  */
 
 import { camelize } from '@ember/string';
-import userGri from 'onedata-gui-websocket-client/utils/user-gri';
 import _ from 'lodash';
 import { A } from '@ember/array';
-import { Promise, resolve } from 'rsvp';
+import { Promise, resolve, all as allFulfilled } from 'rsvp';
 import { get, set } from '@ember/object';
 import groupPrivilegesFlags from 'onedata-gui-websocket-client/utils/group-privileges-flags';
 import spacePrivilegesFlags from 'onedata-gui-websocket-client/utils/space-privileges-flags';
@@ -20,6 +19,7 @@ import harvesterPrivilegesFlags from 'onedata-gui-websocket-client/utils/harvest
 import parseGri from 'onedata-gui-websocket-client/utils/parse-gri';
 import gri from 'onedata-gui-websocket-client/utils/gri';
 import moment from 'moment';
+import { generateSpaceEntityId } from 'onedata-gui-websocket-client/utils/development-model-common';
 
 const USER_ID = 'stub_user_id';
 const USERNAME = 'Stub User';
@@ -84,12 +84,13 @@ const providerClusterDefaultData = {
 export default function generateDevelopmentModel(store) {
   let sharedUsers, groups, spaces, providers, harvesters;
 
-  // create shared users
-  return createSharedUsersRecords(store)
+  return createGuiMessages(store)
+    // create shared users
+    .then(() => createSharedUsersRecords(store))
     .then(su => sharedUsers = su)
     // create main resources lists
     .then(() =>
-      Promise.all(
+      allFulfilled(
         types.map(type =>
           createEntityRecords(store, type, names)
           .then(records => {
@@ -116,9 +117,9 @@ export default function generateDevelopmentModel(store) {
     .then(listRecords => {
       const providers = listRecords[types.indexOf('provider')].get('list');
       const spaces = listRecords[types.indexOf('space')].get('list');
-      return Promise.all([providers, spaces])
+      return allFulfilled([providers, spaces])
         .then(([providerList, spacesList]) =>
-          Promise.all(providerList.map(provider =>
+          allFulfilled(providerList.map(provider =>
             createListRecord(store, 'space', spacesList).then(lr => {
               provider.set('spaceList', lr);
               return provider.save();
@@ -132,9 +133,9 @@ export default function generateDevelopmentModel(store) {
       const providers = listRecords[types.indexOf('provider')].get('list');
       const spaces = listRecords[types.indexOf('space')].get('list');
       const clusters = listRecords[types.indexOf('cluster')].get('list');
-      return Promise.all([providers, spaces, clusters])
+      return allFulfilled([providers, spaces, clusters])
         .then(([providerList, spaceList, clusterList]) =>
-          Promise.all(spaceList.map(space => {
+          allFulfilled(spaceList.map(space => {
             space.set('supportSizes', _.zipObject(
               get(providers, 'content').mapBy('entityId'),
               _.fill(Array(NUMBER_OF_PROVIDERS), perProviderSize)
@@ -144,7 +145,7 @@ export default function generateDevelopmentModel(store) {
               return space.save();
             });
           }))
-          .then(() => Promise.all(clusterList.map(cluster => {
+          .then(() => allFulfilled(clusterList.map(cluster => {
             if (get(cluster, 'type') === 'oneprovider') {
               const clusterProvider =
                 providerList.findBy('entityId', get(cluster, 'entityId'));
@@ -172,8 +173,8 @@ export default function generateDevelopmentModel(store) {
     // add groups, memberships, users and privileges to groups
     .then(listRecords => listRecords[types.indexOf('group')].get('list')
       .then(records =>
-        Promise.all(records.map(record =>
-          Promise.all([
+        allFulfilled(records.map(record =>
+          allFulfilled([
             attachSharedUsersGroupsToModel(
               store, record, 'group', false, sharedUsers, groups
             ),
@@ -188,8 +189,8 @@ export default function generateDevelopmentModel(store) {
       )
       .then(() => listRecords[types.indexOf('space')].get('list')
         .then(records =>
-          Promise.all(records.map(record =>
-            Promise.all([
+          allFulfilled(records.map(record =>
+            allFulfilled([
               attachSharedUsersGroupsToModel(
                 store, record, 'space', false, sharedUsers, groups
               ),
@@ -205,8 +206,8 @@ export default function generateDevelopmentModel(store) {
       )
       .then(() => listRecords[types.indexOf('harvester')].get('list')
         .then(records =>
-          Promise.all(records.map(record =>
-            Promise.all([
+          allFulfilled(records.map(record =>
+            allFulfilled([
               attachSharedUsersGroupsToModel(
                 store, record, 'harvester', false, sharedUsers, groups
               ),
@@ -221,10 +222,10 @@ export default function generateDevelopmentModel(store) {
           ))
         )
       )
-      .then(() => Promise.all(['space', 'group', 'harvester'].map(modelType => {
+      .then(() => allFulfilled(['space', 'group', 'harvester'].map(modelType => {
         return listRecords[types.indexOf(modelType)].get('list')
           .then(records =>
-            Promise.all(records.map(record =>
+            allFulfilled(records.map(record =>
               createPrivilegesForModel(
                 store,
                 record,
@@ -245,6 +246,35 @@ export default function generateDevelopmentModel(store) {
     .then(listRecords => createUserRecord(store, listRecords));
 }
 
+function createGuiMessages(store) {
+  const messages = [{
+    name: 'privacy_policy',
+    body: '<p>Privacy policy</p>',
+  }, {
+    name: 'cookie_consent_notification',
+    body: 'We use cookies for navigation purposes and holding user session state. For more details see our [privacy-policy]privacy policy[/privacy-policy].',
+  }, {
+    name: 'signin_notification',
+    body: 'Onezone will be down for two months. You have 20 seconds to migrate your data. 19... 18...',
+  }];
+
+  return allFulfilled(messages.map(({ name, body }) => {
+    const messageGri = gri({
+      entityType: 'oz_worker',
+      entityId: 'null',
+      aspect: 'gui_message',
+      aspectId: name,
+      scope: 'private',
+    });
+    return store.createRecord('guiMessage', {
+      id: messageGri,
+      gri: messageGri,
+      enabled: true,
+      body,
+    }).save();
+  }));
+}
+
 function createUserRecord(store, listRecords) {
   const spacesIndex = types.indexOf('space');
   return listRecords[spacesIndex].get('list')
@@ -252,7 +282,7 @@ function createUserRecord(store, listRecords) {
     .then(space => space && space.get('entityId'))
     .then(defaultSpaceId => {
       const userRecord = store.createRecord('user', {
-        id: userGri(USER_ID),
+        id: store.userGri(USER_ID),
         fullName: USERNAME,
         basicAuthEnabled: true,
         hasPassword: false,
@@ -283,7 +313,7 @@ function createEntityRecords(store, type, names, additionalInfo) {
     case 'harvester':
       return createHarvesterRecords(store, additionalInfo);
     default:
-      return Promise.all(names.map(number =>
+      return allFulfilled(names.map(number =>
         store.createRecord(type, { name: `${type} ${number}` }).save()
       ));
   }
@@ -323,7 +353,7 @@ function getProviderRecordId(index) {
 }
 
 function createProvidersRecords(store) {
-  return Promise.all(_.range(NUMBER_OF_PROVIDERS).map((index) => {
+  return allFulfilled(_.range(NUMBER_OF_PROVIDERS).map((index) => {
     const providerId = getProviderId(index);
     const id = getProviderRecordId(index);
     const [latitude, longitude] = getCoordinates(index);
@@ -333,7 +363,7 @@ function createProvidersRecords(store) {
       name: PROVIDER_NAMES[index],
       latitude,
       longitude,
-      online: index <= 2,
+      online: index <= 1,
       host: `${providerId}.local-onedata.org`,
     }).save();
   }));
@@ -352,8 +382,14 @@ function generateProviderClusterRecord(index) {
 }
 
 function createSpacesRecords(store) {
-  return Promise.all(_.range(NUMBER_OF_SPACES).map((index) => {
+  return allFulfilled(_.range(NUMBER_OF_SPACES).map((index) => {
     return store.createRecord('space', {
+      id: gri({
+        entityType: 'space',
+        entityId: generateSpaceEntityId(index),
+        aspect: 'instance',
+        scope: 'auto',
+      }),
       name: `Space ${index}`,
       scope: 'private',
       directMembership: true,
@@ -368,13 +404,13 @@ function createSpacesRecords(store) {
 }
 
 function createClientTokensRecords(store) {
-  return Promise.all(_.range(NUMBER_OF_CLIENT_TOKENS).map(() => {
+  return allFulfilled(_.range(NUMBER_OF_CLIENT_TOKENS).map(() => {
     return store.createRecord('clientToken', {}).save();
   }));
 }
 
 function createGroupsRecords(store) {
-  return Promise.all(_.range(NUMBER_OF_GROUPS).map((index) => {
+  return allFulfilled(_.range(NUMBER_OF_GROUPS).map((index) => {
     return store.createRecord('group', {
       name: `group${index}`,
       scope: 'private',
@@ -386,7 +422,7 @@ function createGroupsRecords(store) {
 }
 
 function createLinkedAccount(store) {
-  return Promise.all(LINKED_ACCOUNT_TYPES.map(idp =>
+  return allFulfilled(LINKED_ACCOUNT_TYPES.map(idp =>
     store.createRecord('linkedAccount', {
       idp,
       emails: A([
@@ -399,7 +435,7 @@ function createLinkedAccount(store) {
 
 function createClusterRecords(store) {
   const onezoneId = clusterInstanceGri('onezone');
-  return Promise.all([{
+  return allFulfilled([{
       id: onezoneId,
       gri: onezoneId,
       type: 'onezone',
@@ -427,7 +463,7 @@ function createClusterRecords(store) {
 }
 
 function createHarvesterRecords(store) {
-  return Promise.all(_.range(NUMBER_OF_HARVESTERS).map((index) => {
+  return allFulfilled(_.range(NUMBER_OF_HARVESTERS).map((index) => {
     return store.createRecord('harvester', {
       name: `Harvester ${index}`,
       scope: 'private',
@@ -473,7 +509,7 @@ function createHarvesterRecords(store) {
         },
       }).save().then(() => record)
     ).then(record => {
-      return Promise.all(_.range(3).map((index) => {
+      return allFulfilled(_.range(3).map((index) => {
           return store.createRecord('index', {
             id: gri({
               entityType: 'harvester',
@@ -494,7 +530,7 @@ function createHarvesterRecords(store) {
 }
 
 function createSharedUsersRecords(store) {
-  return Promise.all(_.range(NUMBER_OF_SHARED_USERS).map((index) => {
+  return allFulfilled(_.range(NUMBER_OF_SHARED_USERS).map((index) => {
     return store.createRecord('sharedUser', { name: `sharedUser${index}` }).save();
   }));
 }
@@ -564,7 +600,7 @@ function createPrivilegesForModel(
   groups,
   privilegesFlags
 ) {
-  return Promise.all([
+  return allFulfilled([
     createPrivilegesRecords(
       store, record, modelType, sharedUsers, privilegesFlags, 'user'
     ),
@@ -597,7 +633,7 @@ function createPrivilegesRecords(
   if (modelType === 'group' && privilegesType === 'group') {
     aspect = 'child_privileges';
   }
-  return Promise.all(_.range(sharedGriArray.length).map((index) => {
+  return allFulfilled(_.range(sharedGriArray.length).map((index) => {
     subjectId = parseGri(sharedGriArray[index]).entityId;
     recordData.id = gri({
       entityType: modelType,
@@ -628,14 +664,14 @@ function attachProgressToHarvesterIndices(
   const perHarvesterSeq = 100;
   const maxSeq = perHarvesterSeq * harvestersNumber;
   const lastUpdate = moment().unix();
-  return Promise.all(harvesters.map((harvester, harvesterIndex) => {
+  return allFulfilled(harvesters.map((harvester, harvesterIndex) => {
     const harvesterEntityId = get(harvester, 'entityId');
     return get(harvester, 'indexList')
       .then(indexList => get(indexList, 'list'))
       .then(indices => {
         const indicesNumber = get(indices, 'length');
         const perIndexSeq = Math.ceil(perHarvesterSeq / indicesNumber);
-        return Promise.all(indices.map((index, indexIndex) => {
+        return allFulfilled(indices.map((index, indexIndex) => {
           const indexEntityId = get(index, 'aspectId');
           const currentSeq = harvesterIndex * perHarvesterSeq +
             Math.min((indexIndex + 1) * perIndexSeq, perHarvesterSeq);